BSIS-728: Fix pagination on export donor list
Added settings to the pagination template for the ui-grid plugin
in order to limit the number of pages displayed and use an ellipses
(...).

Unfortunately it was not possible to get rotate to work (so that the
current page is in the middle of the list of numbers), nor boundary
links and numbers. It appears that ui-grid does not support these.

See: http://angular-ui.github.io/bootstrap/#/pagination for a
description of the pagination attributes used in the template.