BSIS-740 Make Array.find polyfill non-enumerable
The find polyfill being added to the Array prototype was enumerable
which meant that looping through the array with for..in would receive
the find function as one of the values. This was causing the test to
fail since an extra permission was being added to the permissions array
with the name of the find function. This function is anonymous so the
value being added was an empty string. By defining the find property as
non-enumerable it will not be included when using for..in and therefore
not included in the permissions so the test passes.