BSIS-888: Add UI to display the stock levels
Added the UI to display stock levels. At the moment the data is
hardcoded.

Previously we used ng-checked to keep the Any/All venue checkbox
automatically checked when the venues ui-select was empty, but I
found issues with error handling. The problem is that ng-model and
ng-checked cannot be used together. The model is not updated when
the checkbox is changed by ng-checked. The model is being used
in order to display error messages for the combination of the
uiselect and checkbox. I implemented an on-change event on the
uiselect box that toggled the allVenues boolean instead of using
ng-checked.

Please see: https://docs.angularjs.org/api/ng/directive/ngChecked