BSIS-887: Hook up Locations by type endpoint
Hook up the endpoint to retrieve a list of DISTRIBUTION_SITE Locations.

Note: At this point the endpoint does not work. It needs the changes
in commit: 2ade8d7f248d8f432fe1e57119749a05058a744b in order to
recognise DISTRIBUTION_SITE as a LocationType in the Repository method
to retrieve the Locations.