BSIS-945 Update grid data after saving
After saving the order form the reference to the items was not being
updated so the grid was not showing the correct data when more items
were added. This change sets the data property of the grid to the
updated order forms items so that more items can be added and are
correctly displayed.