BSIS-969: Add gap count for OrderForm
When a Component is added to the OrderForm, the gap between the units
ordered and supplied needs to be calculate and displayed in the table.

This is done by going through the currently unmatched components when
adding an item to the table and seeing if there is a match. If there
is a match, the gap is decreased by 1 and the units supplied is
increased by 1.