BSIS-965: Hook up find Components endpoint
Hooked up the find Components endpoint using the din and componentCode.

Because ComponentCodes aren't actually numbers (they start with 00) I
had to change the field to a string otherwise the 00 was being lost in
the conversion.

Note: handling the errors will be done in a separate task (BSIS-968)
For example: 404 errors if the component is not found.