BSIS-953: Create HTML page to view & manage an Order Form
Added the HTML and frontend controller used to view an OrderForm.
Linked up the fulfilOrder page, so that when the order is saved, the
user is redirected to the view page. Also if the user presses cancel
on the fulfilOrder page, they will be returned to the view OrderForm
page (without saving the changes).