BSIS-964: Added Find Inventory HTML and frontend controller
Added the HTML and frontend controller for the Find Inventory page.

The sorting of results has been set on the frontend using the sort
attribute in columnDefs and when exporting to PDF, all visible rows
will be exported - which keeps the user defined (or default) sort.

The default sort is DIN and ComponentCode.

The only required fields are DIN if ComponentCode is specified (and
visa versa).