BSIS-964: Change sort on Find Inventory table
Changed the sort on the Find Inventory table to first sort by
Component code, and then DIN. Unfortunately it's not possible
to sort by ComponentType because the ComponentType code is not
a column. And if we sort by ComponentType name, then the sort
order is alphabetical on the name - which seems a little odd
because that column is one of the last columns.