BSIS-964: Fix Find Inventory issue where only visible page was exported
Unfortunately using 'visible' for the PDF export only exports the
currently visible page, so the sorting strategy had to be changed.
(Using visible also exports the sort).

The table sorting has been replaced by an array sort of the response.
The user is able to modify the sort on the page, but that order
will not be exported. This is consistent with the rest of the site.