BSIS-1262 Fix issue when saving donations
The sampleReleased property that was added to the donation in the scope
is sent along to the backend and because it is not a recognised field
the request fails with 400 status. This change does not set
sampleReleased on the donation, but rather just uses the value of
donation.released to determine which value to display.