BSIS-1585: Fix Donation not adding without Adverse Event (#504)
BSIS-1585: Fix Donation not adding without Adverse Event
In the previous correction for BSIS-1530 the $scope.adverseEvent.type
was initialized to a non-null object. This caused the viewDonationBatch.js'
addDonation method to set the donation.adverseEvent even when it was not
set by the end user. This caused an error on the BSIS backend.

This change prevents the viewDonationBatch.js' addDonation
method from setting the donation.adverseEvent if the
$scope.adverseEvent.type is null.