BSIS-1551 Disable pack types based on donation state
If the donation's current pack type does not produce a test sample and
the test batch that the donation belongs to has been released then only
allow other pack types which do not produce a test sample to be
selected. This change makes use of a new endpoint to get the form fields
for editing a donation which includes the status of the test batch to
which the donation belongs if there is one.