BSIS-1681 Check VIEW_TEST_OUTCOME permission when viewing donation
Add a method to AuthService to check if the logged in user has a required permission.
Add a check when viewing donations to ensure that if the logged in user does not have the VIEW_TEST_OUTCOME
permissions the front-end does not attempt to obtain the test outcomes from the back-end.