BSIS-1809: Add required validation to venues
Added required validation to the venues multi-select box on the mobile
clinic export page - so that if the user unchecks the All venues checkbox
they are required to enter a venue in the select box.

I added allVenues to the search model and a function to toggle the
boolean when the user selected or unselected an option from the venue
multi-select box.

I used the uiSelectRequire directive to make the ui-select box required
(there is a bug in ui-select) only when the allVenues flag was false.