BSIS-1822 Fix incorrect premature enabling of "Enter" link for Confirmatory Tests
This is a bug, where if all basicTTITests are entered, the repeat and confirmatory links are always enabled regardless of having pending tests or not.

This was fixed by changing the disabling link condition, to not consider if the basic tests are complete or not, for tti and bloodtyping repeat and confirmatory tests.