BSIS-1809: Redo solution in 4d14df8f5550a68522a0f11e360e5eadccf0a45f
Re implemented an alternative solution for making the selection of venues
required for the Mobile Clinic Export. The use of ng-checked was removed
and replaced with ng-model and the checking and unchecking of the All
Venues checkbox is now handled manually.

Please note that the clear button cannot be of type reset, because
otherwise Angular will handle the clearing of the form and we cannot
use our own defaults.