BSIS-2112 Make componentBatch error message consistent.
The labelling warning/error message shown to user when the component batch
has not been received yet was too specific and therefore inconsistent with
the current warning messages. This commit makes the message shown
when attempting to label a component where its component batch has not been
recieved yet, consistent with the other scenarios.

This commit also makes the warning/error message shown for discarding a
component consistent with the message shown when recording component processed,
in the case where the component batch has not been received yet.