BSIS-2160: Send zip output to dev/null
Send zip output to dev/null because otherwise the build log is too large
and the build fails in Travis.

Error: The log length has exceeded the limit of 4 MB (this usually means
that the test suite is raising the same exception over and over).
The job has been terminated