BSIS-2116 As an implementer, I need BSIS distributables to be included on Github release pages, so that I can implement BSIS without compiling the application (#581)
* BSIS-2160: Zip and publish to release tag

Using Travis to zip and publish the dist folder for BSIS frontend.

https://docs.travis-ci.com/user/deployment/releases/
https://help.github.com/articles/creating-an-access-token-for-command-line-use/

* BSIS-2160: Fix small typo in travis.yml

* BSIS-2160: Get Travis build on branch to test

Added this branch to the branches to run the Travis build for the front
end in order to do testing. Should be reverted once this has been
tested.

* BSIS-2160: Send zip output to dev/null

Send zip output to dev/null because otherwise the build log is too large
and the build fails in Travis.

Error: The log length has exceeded the limit of 4 MB (this usually means
that the test suite is raising the same exception over and over).
The job has been terminated

* BSIS-2160: Remove branches section from travis.yml

Removed the branches section from travis.yml (temporarily) to test if
that is the problem causing the build not to run on the tag.

Seen build message:
Skipping a deployment with the releases provider because this is not a tagged commit

* BSIS-2160 Add debug cli commands in after_success

Add some debug statements to see current directly when executing. I
am not sure if this will work.

* BSIS-2160 Explicitly build the app after success

The dist folder was not being created because the only the tests were being run.

Created a "build" script and called it in the travis file after success.

* BSIS-2160 Explicitly build the app after success second attempt

Tried a different way.

- npm install grunt
- npm install grunt-cli
- grunt build

* BSIS-2160 Update travis.yml to return to parent directory

dist/bsis-frontend.zip cannot be found. This may be due
to the fact that the directory was changed in the after_success
section.

* BSIS-2160 Use TRAVIS_BUILD_DIR in .travis.yml

Using environment variable TRAVIS_BUILD_DIR when deploying
distribution zip to github releases. This is as opposed to
using a short relevant path (dist/bsis-frontend)

* BSIS-2160 Set nodejs version to 4.2.x

Nodejs 4 version was defaulting to 4.x. This
sets nodejs version explicitly to 4.2.x. This
is simply to test to see if it is causing the
current build problem on the node 4 build.

* BSIS-2160 Set node version to 4.2

Node version was 4.2.x. This was incorrect.

* BSIS-2160: Re-organise travis build

Re-organised the Travis build to use the proper events. Also selected
one version of Node so that we can guarantee which version is used
for the release.

* BSIS-2160: Run npm install

* BSIS-2276 Use bsis-release encrypted api_key

Changed api_key value to be that of a bsis-release
github user. It's also now encrypted so that the value
is not in the clear.

* BSIS-2276 Add api_key tag to travis.yml

Previous commit removed the api_key tag in error.

* BSIS-2276 Remove invalid character from .travis.yml

Tab character caused travis.yml file to be invalid. This
commit removes the tab character.

* BSIS-2276 Use unencrypted api_key in travis.yml

Secured encrypted api_key was not working. This is
to test if it was the encryption or the actual token
that is the problem.

* BSIS-2276 Use regenerated encrypted token in travis.yml

Regenerated token in github, and encrypted it via travis
Command line tool.

* BSIS-2279 Update travis to use a script instead of installing grunt-cli

Add grunt-cli to dev dependencies, and use the build script instead of installing grunt-cli for travis release.