BSIS-2287 Revert creation of directive and applying it in the views
The use of the directive to truncate the flag characters, was not going to work in the case that donations with different din lengths existed in the system, so it was decided to revert it.

The full scanned DIN (included the flag characters) will be sent to the backend, and it will be handled there.