BSIS-2289 Fix incorrect reference to donationBatch.createdDate
When editing a testbatch, the donation batch date was not shown in the brackets
next to the venue name. This was because previously the tracking field 'createdDate'
was used in the front-end.

This was replaced with the donationBatchDate and the createdDate is not sent
to the frontend anymore. This commit now references the correct field when
attempting to get the donation batch's date.