BSIS-2309: Fix form clear after label verification
Fixed form clear after label verification so that when scanning paed
components the form is cleared when scanning the second and third
components.

Also removed the clear when changing ComponentType or DIN in the main
search, because it was only a partial clear anyway ... and isn't how
other forms in BSIS work.