BSIS-2370: Use donationDateTime for max times
Using donationDateTime for the check to see if the max time since donation
has been exceeded for a specific component during processing. The createdOn
date of the component might not be the correct date if the parent component
is not the initial component (which is the case for 2nd round processing).