BSIS-2424 Fix aboRhBloodGroups and DonationTypes Reports
This is a merge from support/1.0.x branch, to merge the corrections
implemented in 1.0.1 patch release for the reports.

Previously in the reports, the row values were being overwritten for
each cohort that was instead supposed to be added to the row value.

This commit corrects this by adding the new value to the existing value.