BSIS-2556 Add validation to the Blood Group Select on Find Inventory
Previously when performing an inventory search without the All Blood Groups checkbox
selected, all sites are returned regardless ie. the same result set as Selecting
the All Blood Group checkbox.

This commit adds validation so that if the checkbox is unchecked, at least
one blood group has to be added.