BSIS-2571 Correct "verifying" verify button text
Previously, when the "Verify Pack Label" button is
pressed, the text changes to "Verifying" but did not
revert to "Verify Pack Label" once the response is returned
if it was not verified successfully.

This commit keeps the button disabled in the case that
a failure response is returned, but updates the text. It
also disables the input boxes in this state so that the
user cannot enter the DINs when the Verify button is
disabled.