BSIS-2572 Reset state after label component search
Previously if the user was verifying a label, and
while doing that, pressed search the state would not
be cleared. i.e. the previous verify screen would
still display.

This commit resets the state, so that the user has to
reprint the labels after searching for the new (or the
same) component.