BSIS-2580 Replace Text Input with Date Select & improve Bootstrap usage
Replaced the text input with the date select directive in order to have
a consistent solution accross the site. Then repaired the layout issues
caused by improving the implementation of Bootstrap classes. Layout is
now more responsive and resiliant.

This is being committed seperately from:
https://jembiprojects.jira.com/browse/BSIS-2579

In order to keep commits smaller and more easily reviewed.