BSIS-2537 Read routeParams and set to scope
Previously after searching and copying the link and reloading the
page, the page would reset to it's initial state.

This update Reads route Params for new flaggedForCounselling, Site dropdown
and referred booleans. Also for start and end date. So that the
search details are not last after reloading the link.