BSIS-2694: Add error message for order date
Added an error message for order date on the fulfil Order page because
previously only a red box was displayed to the user if they deleted
the date or edited it and made it invalid.

Note: there is no error message displayed for the Patient birth date.
This is due to an issue with the bootstrap dateselector when it isn't
required the "date" error is not assigned to the component.