BSIS-2635 Fix regex pattern in header.js path
Previously header.js path was incorrect for manageTransfusionReactionType.
It was configured in such a way to only work when there was an integer
after the manageTransfusionReactionType/<x> where <x> is the ID of the
TransfusionReactionType.

This commit makes it work for both scenarios 1) when adding a new
TransfusionReactionType (without an ID) and 2) when updating
an existing TransfusionReactionType (with an ID)