BSIS-2661 Refactor check of componentCode in componentTypes
This commit updates the filter that checks if a componentCode (in the transfusion record) is in the list of componentTypes.
The new filter condition checks if componentType.componentTypeCode is contained in transfusion.componentCode (instead of checking for equality); this change allows taking into consideration the case when componentCode is different from componentTypeCOde (ex. 2101-01 2101-02).