BSIS-2727: Fix directives when modelValue is null
Fixed date/time directives so that if the modelValue is null or undefined
the directive returns true (bypassing the validation).

This is important for when the input field is not mandatory and doesn't
have any value entered. This happens on "Add Transfusion" when entering
the patient's date of birth.

A fix was also done for:
 - time-not-same-as (Bleed time on Record Components)
 - time-is-after (Processing time on Record Components)
 - ui-date-range (To/From period searches e.g. Find Components)