BSIS-2834 Use transfusionSiteId in updateAllSites
The variable $scope.search.usageSite has been replaced with $scope.search.transfusionSiteId (to match the change of the back-end query parameter). This commit changes updateAllSites to use the new variable name which fixes a bug ("All site" check-box was not unchecked when the user selects a site from the drop-down).