BSIS-2720 Add methods to report generator service
Added methods generateDataRowsGroupingByCohort and convertRowObjectsToArrays.

generateDataRowsGroupingByCohort generates rows by cohort for a given cohort category, doesn't take location into account. Can be used in a report and also in a summary.

convertRowObjectsToArrays is used when using a list of row objects in the pdf. Ui grid does this conversion automatically, but to calculate a summary and show it in the pdf, for example, we need to do the conversion manually.

Row object = row.location, row.componentType, row.gap
Row array = [location, componentType, gap]

Also, added a check for when the cohort category is not found in the dataValue.