BSIS-2848: Fix issue with display of unit gap
Fixed issue where the backend was returning ComponentViewModels in the
OrderFormViewModel, but we were retrieving InventoryViewModels when
the user added a Component to an OrderForm. This mean that the items in
the table were a mix of different view models. This caused issues because
in InventoryViewModel there is an attribute bloodGroup, but in ComponentViewModel
it is two attributes - bloodAbo and bloodRh. The backend now uses
InventoryViewModels for OrderFormViewModel components.