BSIS-2954 Fix 404 error when recording empty component code
When the component code is manually deleted, the corresponding model ($scope.transfusion.componentCode) is set to empty string (i.e. ""). The back-end expects null when no component code is provided; when an empty string is provided instead, a 404 error is returned. This commit fixes the problem by explicitly setting $scope.transfusion.componentCode to null when it has an empty field.