BSIS-3109: Add translations.html
Added translations.html which will contain text that must be translated
but that is not being picked up by the Grunt ng-gettext extract process.
The extract process cannot pick up text that is generated in the backend
(e.g. enums) and is not properly picking up some filtered text (e.g. the
dynamic buttons and filtered enum values).

Because we cannot edit the template.pot file because it is automatically
generated by the "extract" process and because we cannot manually add text
to the po files, having an extra HTML file containing translateable elements
allows us to force hidden text into the translation process.

It unfortunately will be a manual process to ensure that all hidden text
is referenced in this file. As a rule of thumb, all enums should be added
and any filter with a complex expression (one containing a variable).