BSIS-3067: Add translations for viewTestBatch
Added translations for viewTestBatch - for the text in the HTML, the UI
grid, the enums, the popups and the PDF export.

Removed generic server side error handling as this hasn't been implemented
on other pages and it's unclear how we will be translating that.

Actual translations and entries into translations.html will be completed
in a separate task.