BSIS-3187: Move <hr> tag out of <table>
Removed the <hr> tag from inside the table and put it under the heading,
where it should be shown. The <hr> tag was causing inconsistant rendering
of the HTML which was causing gettext some issues - on the test server
the <hr> element was being rendered inside the <th> for "Tested On".