BSIS-3150: i18n for mergeDonorsDonations
Added i18n for Merge Donors Donations and fixed an issue where a configurable
property (weight unit) was hardcoded. Note that usually with an abbreviation
such as 'Wt' a translation comment should be added, but there doesn't seem
to be able to be a way to do so with filters. Only in the JS and with the
directive.