BSIS-3155: finish i18n for manageDuplicateDonors
Finished i18n for Manage Duplicate Donors wizard - gender needed to be
translated and the buttons were in a div that was too small which is
a problem if the translated text is longer than the English text. It's
also unnecessary to add spans for buttons - just delete the whitespace
in the button tag.