BSIS-3162 Translate 'to' between startTime and endTime
To was not being translated. This commit allows the
to be translated.

The start and end time has special attributes because the donation
can be edited on this screen. This prevented the use of translate-params
to make the 'startTime to endTime' translatable as a whole sentence
as was done for BSIS-3158