BSIS-3162 Use generic i18n message for donation properties
Use generic error message for "Please enter value below x"
and "Please enter value below y" so that this only has to be
translated once.

Also correct error in translate-params attribute. It was previously
translation-param which was not correct. It also was not following
camel case convetion and therefore the scope variable was being
picked up instead of the actual translate param