BSIS-3215: Add translation comments for enums
Added translation comments for enums in translations.html to indicate
that the string to translate is part of a group of translations, which
can help the translator.

Also checked the references of the enum values in the rest of the app
(right click on the string in POEdit to see those) to ensure the term
wasn't being used in a different context elsewhere. If the string was
being used to describe the enum value (e.g. as a table heading or
as a radio button) no other comment was added, but if the string was
used elsewhere to mean something else (e.g. Unknown) then the other
reference was given a comment so the translator knows it is being used
to mean two different things.