BSIS-3268: Fix incorrect usage of <br/>
Fixed the incorrect usage of the <br/> tag which was causing unnnecessary
line breaks in the field when there wasn't an error message displayed. Line
break tags are generally unnecessary - in this case using a <div> instead
of a <span> causes a newline that achieves the same effect when displaying
error messages without the side effect when no error messages are being
displayed.