BSIS-3276 Fix bug: table headers not being translated
Adding translation to displayName fields (in $scope.gridOptions.columnDefs) did reflect on the UI (i.e column rendered without the translation). The issue was that when $scope.gridOptions parameters are set, the column definitions are not available yet (because this report retrieves some of the columns from the back-end). When the column definition is empty (i.e. [], the default behavior of gridOptions is to use the keys of data arrays as column header, this is why this issue was not picked-up earlier (i.e. the field keys used happened to match the column headers).

To fix the above issue, this commit wraps $scope.gridOptions = {...} in a function (initGridOptionsConfig) which is called when the back-end returns the columns which ensures that $scope.gridOptions.columnDefs is not empty.