BSIS-3309: Fix footer Total {{recordsName}}
Fixed the translation of recordsName which wouldn't have been translated
if it was a variable (only the text is translated, not the variable).

The text was translated in the controller (and not the ReportsLayoutService)
which has resulted in a little more code, but this approach is more
straightforward. It is clear by reading the code in the Controller that
the text has been translated and has not been forgotten.