BSIS-3288 Revert changes to referred and status el
Reverted changes to the referred and counselling status form elements
because the design is different to other pages in the BSIS app (wrt
to using toggle boxes instead of check boxes for booleans and using
toggle boxes instead of drop downs for lists). This will be addressed
in a system-wide strategy instead.