BSIS-3432: Fix invalid HTML and standardise l&f
Fixed the invalid HTML (where there is an <hr> inside a <table>) which was
causing different HTML to be rendered on a compiled version of BSIS. The
different HTML resulted in the <hr> being inserted into a <th> tag that was
translated which meant that the text inside the <th> tag could not be
matched in translations.js because it contained <hr> html.

Also standardised the other screens that display test outcomes to use
the <hr> under the table heading (not the table columns).

Test outcomes for a donation are displayed on these pages:
 - View Donor Dashboard under Donations tab, click "Show Test Outcomes"
 - Manage Donation Batch, click the Donation and then "Show Test Outcomes"
 - Donor Counselling, search and click a Donor in the search results
 - View Test Sample, search for a DIN