BSIS-3422: Fix translation of ng-messages
Fixed translation of ng-messages where the translate directive was put into the
same HTML element containing the ng-message directive. This doesn't work
because the ng-message directive takes precidence and the translate one is
ignored, so I added a separate <translate> around the text to translate.

Note: usually we've been adding a separate span, but actually it's not
technically necessary to do that.